+++ src/main/java/com/folks/app/bo/ProfessionalServiceBO.java
@@ -96,7 +96,7 @@ public class ProfessionalServiceBO {
         return rows;
     }
 
-    public ProfessionalService view(AppUser usr, Long id) {
+    public ProfessionalService view(AppUser usr, Integer id) {
         StopWatch timer = StopWatch.newTimer();
         timer.start();
 
@@ -111,7 +111,7 @@ public class ProfessionalServiceBO {
         return professionalService;
     }
 
-    public ProfessionalService remove(AppUser usr, Long id) {
+    public ProfessionalService remove(AppUser usr, Integer id) {
         StopWatch timer = StopWatch.newTimer();
         timer.start();
